create idea_agg to idea_vld tests scaffolding
